Refactor to purely functional interface
Fully commit to a functional interface for the abstract methods
which the user needs to implement. Before, some methods like "fit"
would return the model state, while others like "chain" would set
an object attribute to store the model state. Now, all state is
always returned (functional), and we have private wrapper methods
to handle setting object attributes.

This was done to be more consistent, so that expectations and
interfaces are clearer for users.